Backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pdfplumber
python-docx
requests
python-dotenv
pytest
httpx
pygraphviz
langgraph